+++ Sources/Core/Debug.swift
@@ -1,5 +1,11 @@
 import Foundation
 
+// for flags in swift packages: https://stackoverflow.com/questions/38813906/swift-how-to-use-preprocessor-flags-like-if-debug-to-implement-api-keys
+//swiftSettings: [
+//    .define("VAPOR")
+//]
+
+
 @available(*, deprecated, renamed: "CustomError")
 public typealias KuError = CustomError 
 
@@ -32,25 +38,46 @@ public enum CustomError: Error {
     }
 }
 
-public enum DebugLevel: Comparable, CustomStringConvertible {
+
+public enum DebugLevel: Comparable, CustomStringConvertible, CaseIterable {
+    /// Only use .OFF for setting the current debug level so nothing is printed.  If you wish to disable a debug message, use .SILENT
     case OFF
     case ERROR
     case WARNING
     case NOTICE
     case DEBUG
+    case SILENT
     /// Change this value in production to DebugLevvel.ERROR to minimize logging.
-    #if DEBUG
+    // set default debugging level to .DEBUG (use manual controls to turn OFF if not debug during app tracking since previews do not have app tracking set up nor does it have compiler flags or app init.
     public static var currentLevel = DebugLevel.DEBUG
-    #else
-    public static var currentLevel = DebugLevel.NOTICE
-    #endif
+    
     public static var defaultLevel = DebugLevel.ERROR
+    
+    /// Set this to `true` to log failed color parsing notices when returning `nil`
+    public static var colorLogging = false
+    
     /// setting this to false will make debug( act exactly like print(
     public static var includeContext = true
+    public var emoji: String {
+        switch self {
+        case .OFF:
+            return "💤"
+        case .ERROR:
+            return "🛑"
+        case .WARNING:
+            return "⚠️"
+        case .NOTICE:
+            return "ℹ️"
+        case .DEBUG:
+            return "🪲"
+        case .SILENT: // should not typically be used
+            return "🤫"
+        }
+    }
     public var symbol: String {
         switch self {
         case .OFF:
-            return ""
+            return "-"
         case .ERROR:
             return "•"
         case .WARNING:
@@ -59,6 +86,25 @@ public enum DebugLevel: Comparable, CustomStringConvertible {
             return ">"
         case .DEBUG:
             return ":"
+        case .SILENT: // should not typically be used
+            return "z"
+        }
+    }
+    public var color: any KuColor {
+        let type = KuColor.DefaultColorType.self
+        switch self {
+        case .OFF:
+            return type.gray
+        case .ERROR:
+            return type.red
+        case .WARNING:
+            return type.yellow
+        case .NOTICE:
+            return type.blue
+        case .DEBUG:
+            return type.green
+        case .SILENT: // should not typically be used
+            return type.black
         }
     }
     public var description: String {
@@ -73,6 +119,8 @@ public enum DebugLevel: Comparable, CustomStringConvertible {
             return "NOTICE"
         case .DEBUG:
             return "DEBUG"
+        case .SILENT:
+            return "SILENT"
         }
     }
     /// use to detect if the current level is at least the level.  So if the current level is .NOTICE, .isAtLeast(.ERROR) = true but .isAtLeast(.DEBUG) = false.  Will typically be used like: if DebugLevel.currentLevel.isAtLeast(.DEBUG) to check for whether debugging output is on.  Simplify using convenience static func DebugLevel.isAtLeast(.DEBUG) 
@@ -82,6 +130,22 @@ public enum DebugLevel: Comparable, CustomStringConvertible {
     /// use to detect if the current level is at least the level.  So if the current level is .NOTICE, .isAtLeast(.ERROR) = true but .isAtLeast(.DEBUG) = false.  Will typically be used like: if DebugLevel.isAtLeast(.DEBUG) to check for whether debugging output is on. 
     public static func isAtLeast(_ level: DebugLevel) -> Bool {
         return Self.currentLevel.isAtLeast(level)
+    }
+    
+    mutating func rotate() {
+        let cases = DebugLevel.allCases
+        guard var index = cases.firstIndex(of: self) else {
+            fatalError("Unknown case!")
+        }
+        index++
+        if index >= cases.count {
+            index = 0
+        }
+        // can't print or access the currentLevel while mutating the same value
+//        print("Current debug level: \(DebugLevel.currentLevel.description)")
+        //print("Rotating to: \(index): \(cases[index].description)")
+        self = cases[index]
+//        print("Current debug level: \(DebugLevel.currentLevel.description)")
     }
 }
 //DebugLevel.currentLevel = .ERROR
@@ -103,7 +167,7 @@ public func debug(_ message: Any, level: DebugLevel = DebugLevel.defaultLevel, f
     let simplerFunction = function.replacingOccurrences(of: "__preview__", with: "_p_")
     let threadInfo = Thread.isMainThread ? "" : "^"
     if DebugLevel.includeContext {
-        print("\(simplerFile)(\(line)) : \(simplerFunction)\(threadInfo)\n\(level.symbol) \(message)")
+        print("\(simplerFile)(\(line)) : \(simplerFunction)\(threadInfo)\n\(level.emoji) \(message)")
     } else {
         print(message)
     }